requirements.txt
beautifulsoup4>=4.12.3
dashscope>=1.23.4
feedparser>=6.0.11
fastapi>=0.115.12
jinja2>=3.1.6
Pillow>=11.2.1
python-dateutil>=2.9.0.post0
python-docx>=1.1.2
python-dotenv>=1.0.1
PyYAML>=6.0.2
requests>=2.32.3
schedule>=1.2.2
uvicorn>=0.34.0
